Add readRpcConfig fn, move other rpc config fns to new file
