--- src/features/pastelConf/masternodes.ts
@@ -1,250 +1,0 @@
-export default [
-  {
-    label: 'addnode',
-    value: '104.207.135.40',
-  },
-  {
-    label: 'addnode',
-    value: '104.238.133.186',
-  },
-  {
-    label: 'addnode',
-    value: '104.238.162.40',
-  },
-  {
-    label: 'addnode',
-    value: '128.199.187.212',
-  },
-  {
-    label: 'addnode',
-    value: '128.199.188.24',
-  },
-  {
-    label: 'addnode',
-    value: '13.59.218.7',
-  },
-  {
-    label: 'addnode',
-    value: '136.244.86.129',
-  },
-  {
-    label: 'addnode',
-    value: '139.180.168.9',
-  },
-  {
-    label: 'addnode',
-    value: '140.82.25.45',
-  },
-  {
-    label: 'addnode',
-    value: '142.93.151.71',
-  },
-  {
-    label: 'addnode',
-    value: '143.198.51.103',
-  },
-  {
-    label: 'addnode',
-    value: '143.198.51.61',
-  },
-  {
-    label: 'addnode',
-    value: '143.198.54.53',
-  },
-  {
-    label: 'addnode',
-    value: '143.198.57.203',
-  },
-  {
-    label: 'addnode',
-    value: '143.198.58.140',
-  },
-  {
-    label: 'addnode',
-    value: '143.198.59.3',
-  },
-  {
-    label: 'addnode',
-    value: '144.202.35.14',
-  },
-  {
-    label: 'addnode',
-    value: '149.248.21.43',
-  },
-  {
-    label: 'addnode',
-    value: '149.28.133.64',
-  },
-  {
-    label: 'addnode',
-    value: '149.28.250.189',
-  },
-  {
-    label: 'addnode',
-    value: '149.28.36.126',
-  },
-  {
-    label: 'addnode',
-    value: '149.28.51.31',
-  },
-  {
-    label: 'addnode',
-    value: '149.28.95.51',
-  },
-  {
-    label: 'addnode',
-    value: '155.138.130.145',
-  },
-  {
-    label: 'addnode',
-    value: '155.138.165.93',
-  },
-  {
-    label: 'addnode',
-    value: '157.230.115.155',
-  },
-  {
-    label: 'addnode',
-    value: '158.247.226.157',
-  },
-  {
-    label: 'addnode',
-    value: '159.65.149.140',
-  },
-  {
-    label: 'addnode',
-    value: '159.65.198.238',
-  },
-  {
-    label: 'addnode',
-    value: '159.65.202.49',
-  },
-  {
-    label: 'addnode',
-    value: '159.89.164.214',
-  },
-  {
-    label: 'addnode',
-    value: '165.227.130.203',
-  },
-  {
-    label: 'addnode',
-    value: '165.227.161.206',
-  },
-  {
-    label: 'addnode',
-    value: '165.227.169.213',
-  },
-  {
-    label: 'addnode',
-    value: '165.232.138.76',
-  },
-  {
-    label: 'addnode',
-    value: '18.188.63.158',
-  },
-  {
-    label: 'addnode',
-    value: '18.217.242.134',
-  },
-  {
-    label: 'addnode',
-    value: '18.222.110.225',
-  },
-  {
-    label: 'addnode',
-    value: '18.223.107.141',
-  },
-  {
-    label: 'addnode',
-    value: '188.166.101.202',
-  },
-  {
-    label: 'addnode',
-    value: '188.166.114.111',
-  },
-  {
-    label: 'addnode',
-    value: '188.166.160.37',
-  },
-  {
-    label: 'addnode',
-    value: '192.248.191.36',
-  },
-  {
-    label: 'addnode',
-    value: '207.246.127.4',
-  },
-  {
-    label: 'addnode',
-    value: '3.129.210.72',
-  },
-  {
-    label: 'addnode',
-    value: '3.129.68.144',
-  },
-  {
-    label: 'addnode',
-    value: '3.135.238.251',
-  },
-  {
-    label: 'addnode',
-    value: '3.16.27.182',
-  },
-  {
-    label: 'addnode',
-    value: '3.21.190.142',
-  },
-  {
-    label: 'addnode',
-    value: '3.21.242.7',
-  },
-  {
-    label: 'addnode',
-    value: '45.63.11.40',
-  },
-  {
-    label: 'addnode',
-    value: '45.76.147.34',
-  },
-  {
-    label: 'addnode',
-    value: '45.76.5.15',
-  },
-  {
-    label: 'addnode',
-    value: '45.77.221.144',
-  },
-  {
-    label: 'addnode',
-    value: '52.15.245.70',
-  },
-  {
-    label: 'addnode',
-    value: '68.183.192.150',
-  },
-  {
-    label: 'addnode',
-    value: '68.183.193.6',
-  },
-  {
-    label: 'addnode',
-    value: '68.183.194.2',
-  },
-  {
-    label: 'addnode',
-    value: '68.183.204.117',
-  },
-  {
-    label: 'addnode',
-    value: '78.141.217.183',
-  },
-  {
-    label: 'addnode',
-    value: '95.179.188.82',
-  },
-  {
-    label: 'addnode',
-    value: '95.179.214.156',
-  },
-]
